// src/app/models/locations.ts
export interface LocationDetails {
  levelName: string;
  levelCode: string;
  children: LocationDetails[];
}




// export interface LocationDetails {
//   leaf_Code: string
//   leaf_Name: string
//   rollup_Level_7_Code: string
//   rollup_Level_7_Name: string
//   rollup_Level_6_Code: string
//   rollup_Level_6_Name: string
//   rollup_Level_5_Code: string
//   rollup_Level_5_Name: string
//   rollup_Level_4_Code: string
//   rollup_Level_4_Name: string
//   rollup_Level_3_Code: string
//   rollup_Level_3_Name: string
//   rollup_Level_2_Code: string
//   rollup_Level_2_Name: string
//   rollup_Level_1_Code: string
//   rollup_Level_1_Name: string
//   rollup_Operato: any
//   report_Categor: any
//   custom_Members: any
//   branch_City: any
//   st: any
//   legal_Entity: any
//   legacy_Company: any
//   branch_Type: any
//   country: any
//   branch_Number: any
//   ssS_Type: any
//   region: any
//   branch_Source: any
//   createD_BY: any
//   createD_TIMESTAMP: string
//   }